cobc: * codegen.c (output_initialize_compound): adjusted OCCURS initialization to only do the initialization (and runtime checks) once and then memcpy to all other occurences * tree.c: preparation for all-upper-case names (post 3.1)
libcob:
* fileio.c: initialize errno in all places where it is checked afterwards
* fileio.c [_WIN32]: implemented file locking via LockFileEx/UnlockFile
* common.c (cob_sys_waitpid) [_WIN32]: fixed logic error in #if, allowing the best process synchronization possible on that platform
* common.h [_MSC_VER]: removed unused global includes

tests:
* atlocal.in, atlocal_valgrind, atlocal_win:
  * fix unsetting of variables which values contain "COB" (rendered the testuite unusable before...)
  * allow screenio-tests to be run on cygwin/msys with ncurses out-of-the box
  * fixed TEST_LOCAL (may not use pre-inst-env) and "external" versions with old indexed file msgid
  * worked around some win32 issues

+++ libcob/screenio.c
@@ -1355,7 +1355,7 @@ find_field_by_pos (const int initial_curs, const int line, const int column) {
 		if (line == sline
 		 && column >= scolumn
 		 && column <= right_pos) {
-			return idx;
+			return (int)idx;
 		}
 	}
 	return -1;
